Re-organizing the types to include new optional data change
