+++ src/eosjs-api.ts
@@ -18,7 +18,6 @@ import {
     SignatureProvider,
     TransactConfig,
     Transaction,
-    TransactResult,
 } from './eosjs-api-interfaces';
 import { JsonRpc } from './eosjs-jsonrpc';
 import {
@@ -28,7 +27,8 @@ import {
     PushTransactionArgs,
     GetBlockHeaderStateResult,
     GetBlockInfoResult,
-    GetBlockResult
+    GetBlockResult,
+    TransactResult,
 } from './eosjs-rpc-interfaces';
 import * as ser from './eosjs-serialize';
